Fix preloader issues with srcdoc documents
PreloadRequest was figuring out the full URL from the document's url()
when <base href> was not present in the document.

That was creating an issue in cases where url() is not the same as the
document's m_baseURL. One such case is srcdoc based documents, where
url() is "about:srcdoc" whereas m_baseURL is the embedding document.

The result of that was that documents inside srcdoc were not getting
their resources preloaded unless they contained a <base href>.

This CL fixes that.

BUG=564536

Review URL: https://codereview.chromium.org/1487373002

Cr-Commit-Position: refs/heads/master@{#362716}