Use If-Range instead of If-Match/If-Unmodified-Since for partial requests.
If-Range can shortcut partial resumption requests by having the server
respond with the entire entity if there's a validator mismatch. This
cuts down resumptions by one roundtrip.

The If-Range has been in use by HttpCacheTransaction successfully and it
is not believed that availability of support is a concern.

R=rdsmith@chromium.org,svaldez@chromium.org,mkearney@chromium.org,davidben@chromium.org,isherman@chromium.org
BUG=7648

Review URL: https://codereview.chromium.org/1444253003

Cr-Commit-Position: refs/heads/master@{#362771}