Refactor SVGAnimatedLength negative values mode
The SVGAnimatedLength constructor takes an additional argument (compared to the
arguments passed to other SVGAnimatedProperty classes) which represents whether
the length may be negative. Although this is stored as a property of
SVGAnimatedLength, the value is determined by the particular SVG attribute that
the SVGAnimatedLength is representing, and does not change.

In order to facilitate the implementation of additive animations for
AnimatedLength, this patch removes this additional argument, and instead
implements a helper function in SVGLength.h to look up the value for a given SVG
attribute.

BUG=530436

Review URL: https://codereview.chromium.org/1491023003

Cr-Commit-Position: refs/heads/master@{#362941}