[css-grid] Fix items' height computation in indefinite grids with fr
Since master@{#358816} we run a second round of the track sizing algorithm
whenever the height of the grid is indefinite in order to properly compute
row sizes. The available space passed to the track sizing algorithm must not
contain neither borders nor paddings, otherwise it will think that it has
more space available than the existing one. We should use the height of the
content box instead.

BUG=560771,562985

Review URL: https://codereview.chromium.org/1490093003

Cr-Commit-Position: refs/heads/master@{#362957}