Add border-image longhands to border shorthand
border-image longhands should be treated as reset-only sub properties
of border, i.e. setting border resets border-image to it's initial
value EXCEPT when border is set to a CSS-wide keyword like
'initial', in which case the border-image longhands are also set to that value).

This is explicitly stated in specs in these 2 places:
https://drafts.csswg.org/css-cascade/#reset-only-sub-property
https://drafts.csswg.org/css-backgrounds-3/#the-border-shorthands

Firefox inherits correctly in this way
IE does not inherit correctly

BUG=557159

Review URL: https://codereview.chromium.org/1492663004

Cr-Commit-Position: refs/heads/master@{#363087}