OverlayProcessor subtracts damage by underlay planes if possible
Previously it wasn't possible for the OverlayProcessor to change the
damage rect when it scheduled something as an underlay. Now it saves the
visible rect each time an underlay is scheduled. If the next frame also
has an underlay it checks whether:
  a) the underlay visible rect matches that of the previous frame, and
  b) there are no visible quads on top of it.
When both are true, the framebuffer will have a hole in the right place,
and doesn't need to be redrawn. It's important to note that this relies
on the OverlayProcess knowing about each frame, so it can reason about
the state of the framebuffer. For this, a new SkipProcessForOverlays()
lets the Processor know that it should reset the saved visible rect.

BUG=533630
TEST=cc_unittests
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1434913003

Cr-Commit-Position: refs/heads/master@{#363306}