Make use of channel setup/teardown functions symmetric in ChildProcessHostImpl.
In some edge cases a ChildProcessHostImpl can be created but never be given a channel. As currently written the RegisterCommunicationChannel and OnFilterAdded functions aren't called until the channel is added. However, the DeregisterCommunicationChannel and OnFilterRemoved functions are always called. In filters that expect calls to these functions to come in pairs this can lead to problems.

This CL makes it so that the equivalent channel teardown functions are only invoked if a channel was actually created. This is the same way RenderProcessHostImpl handles things.

BUG=556584

Review URL: https://codereview.chromium.org/1501733002

Cr-Commit-Position: refs/heads/master@{#363572}