[sql] Remove part of WebDatabase SQLite patch.
Due to how POSIX advisory locking is specified (see "Posix Advisory
Locking" in SQLite's os_unix.c), SQLite needs to maintain an additional
ref-counting structure.  The routines to get, update, and destroy
"reusable_file_handle" were originally exposed to allow WebDatabase to
interact with that solution on POSIX.  findReusableFd() calls stat() on
the filename to determine the inode and device to figure out if a
pending-close file descriptor can be reused.  Since WebDatabase file
names are resolved by the browser, this stat() can never succeed in the
renderer, thus WebDatabase can never re-used a pending-close fd.

Removing the findReusableFd() wrapper means that the update and destroy
wrappers are also no longer relevant, and the entire interface is
reasonable to consolidate into the initialize call.  This reduces code
sharing, but since the code in question was never designed to be shared
this way, not much is lost.

Re-using reusable file descriptors is an optimization to reduce
close/open churn, but not necessary to correct operation.  Keeping the
set of closed-while-locked descriptors is necessary to correct
operation, and this CL leaves SQLite to track and handle that.

[Rewrites https://codereview.chromium.org/1473963002 which was reverted
 because it caused crashes.]

BUG=563319,562440

Review URL: https://codereview.chromium.org/1483273002

Cr-Commit-Position: refs/heads/master@{#363816}