Fixing a crash that occurs when our feedback animation is interrupted.
Our input feedback animation code handles being interrupted, however if
the timing is just right, the duration since the start of the animation
will cause the radius of the animation to be 0 which is an invalid
parameter for the RadialGradient class we use to draw the feedback.  My
fix is to prevent this from occurring by exiting the render() function
early and allowing the animation to begin on the subsequent render()
call.

BUG=

Review URL: https://codereview.chromium.org/1506013003

Cr-Commit-Position: refs/heads/master@{#363851}