mus: Use ui::Event in DisplayManagerDelegate.
The window server in mus needs to create a mojom::Event out of a ui::Event before
it can be dispatched to the client. However, internally, the window server can
still use ui::Events it receives from the PlatformWindow. This patch delays the
conversion from ui::Event to mojom::Event a little, and so DisplayManager can
send a ui::Event to its delegate.

BUG=none

Review URL: https://codereview.chromium.org/1512003002

Cr-Commit-Position: refs/heads/master@{#364161}