Adding a flag for drawing the cursor to the RenderData class.
With the introduction of new input modes, we need to be able to figure
out whether the current input strategy wants the local cursor bitmap
to be rendered or not.

I have added this attribute to the RenderData class which is set by the
InputStrategy when it is instantiated and read by the DesktopView during
each render loop.

This change was pulled from a larger CL.  The original impl had this
flag in the DesktopView class instead with a suggestion of making it
volatile or an AtomicBoolean.  This flag will be changed a handful of
times a session but read many thousands of times.  Therefore I did not
want to add overhead to the read operation that would come from the
volatile/AtomicBoolean approach.  Instead I want to use the mRenderData
sync mechanism (which is already acquired within the render loop) so no
overhead is added.

BUG=

Review URL: https://codereview.chromium.org/1514453006

Cr-Commit-Position: refs/heads/master@{#364182}