Fix crash due to delayed account changed notification.
Seeding account id <-> name mapping and ValidateAccountSettings(...)
(which fires token refreshed/revoked) notifications independently query
the OS for the current set of accounts.

As of http://crrev.com/1470263002, we notify the seeder immediately on
account changes so that ValidateAccountSettings can be deferred. That is
still not enough (as per http://crash/ab7b830f696418f4).

Part of the issue is there isn't a canonical source within Chrome for
the accounts (AccountTrackerService is seeded the id<->name mapping and
never removed so it may be stale). Thus a preferred solution (IMHO) would
involve tightly coupling the two actions (seeding and validation) thus acting
on the same getAccounts query to the OS and populating the ATS only after that.

I haven't completely understood the implications of this and would rather not
go on a major yak shave (I'm OOO till next year).
I would appreciate your thoughts on resolving this properly, till then
this patch ignores such errors instead of crashing, which is actually fine
practically.

BUG=535211

Review URL: https://codereview.chromium.org/1509453002

Cr-Commit-Position: refs/heads/master@{#364312}