MediaRecorder test: disable flaky time stamp checks.
MediaRecorder testing start(time_slice): disable the flaky
time stamps delta calculations because they can be slightly
off and cause some bots to fail. The root cause is the time
stamp delta was supposed to be 100ms but it came out as
99.99999999999955ms which failed (sic.). The corrective
action is to remove that time stamp check.

Removal of this check is fine because the checks pertains
to a performance measurement, which is not necessary in a
content browser test.

The flakiness had only been found on MAC bots so far.

BUG=568913

mcasas@ PTAL
avi@ Owners RS

Review URL: https://codereview.chromium.org/1515323002

Cr-Commit-Position: refs/heads/master@{#364859}