Don't defproc the WM_NCLBUTTONDOWN message with HTCAPTION as the hittest code.
The default windows behavior here is to post the WM_SYSCOMMAND message with SC_MOVE. While that is desired
there are cases like the user clicking the caption and not moving the mouse. In this case the WM_SYSCOMMAND
defproc handling goes into an initial modal loop presumably calling DragEnter to see if a drag is occurring.
That loop does not dispatch all messages and instead appars to be peeking for mouse input only. This causes
our UI message pump wake up message to not get processed leading to stuff on the window not painting etc.

Proposed fix is to not defproc the WM_NCLBUTTONDOWN message and instead wait for a WM_NCMOUSEMOVE message to
a different location to actually defproc the message. We keep track of whether the mouse down was received
and its location.

BUG=486169
TEST=Covered by views unittest WidgetTest.SysCommandMoveOnNCLButtonDownAndMoveTest

Review URL: https://codereview.chromium.org/1504333013

Cr-Commit-Position: refs/heads/master@{#364892}