Fix races involving detaching and navigating a remote frame.
With OOPIF, we were hitting a couple of crashes in RenderFrameImpl on
the remote-to-local navigation path, due to not finding a proxy to
replace.  These crashes were caused by races between a remote-to-local
navigation on a proxy and the parent detaching that proxy.  This CL
provides fixes for three such races.

1) Proxy is detached after the browser process initiates the
remote-to-local navigation, but before the renderer process receives
the FrameMsg_NewFrame that creates the provisional frame.  We were
hitting a CHECK(proxy) in RenderFrameImpl::CreateFrame in that case,
since the proxy with the provided routing ID was already deleted.
This CL adds a check to not create the local frame if the proxy it is
going to replace is already gone.
Test: NavigateProxyAndDetachBeforeProvisionalFrameCreation

2) Proxy is detached after RenderFrameImpl::CreateFrame creates the
provisional local frame, but before FrameMsg_OnNavigate is received.
Previously, we proceeded to initiate the navigation, which eventually
committed and hit a CHECK(proxy) when trying to find the proxy to swap
with.  This CL fixes this case to not start the navigation in the
first place, since the frame being navigated is already gone.
Test: RenderViewImplTest.NavigateProxyAndDetachBeforeOnNavigate
(with --site-per-process)

3) Proxy is detached after FrameMsg_OnNavigate starts the provisional
load, but before it commits.  This would also hit a CHECK(proxy) in
didCommitProvisionalLoad when trying to find the proxy to swap with.
This CL adds a check for this case in didCommitProvisionalLoad.
Test: NavigateProxyAndDetachBeforeCommit

In cases 2) and 3), the provisional local frame should also be cleaned
up.  This is not currently done, but Nasko is adding that logic as
part of fixing https://crbug.com/548275, where if a parent detaches a
remote child, the browser process will send IPCs to clean up
corresponding local frames.

BUG=526304, 568676

Review URL: https://codereview.chromium.org/1515243002

Cr-Commit-Position: refs/heads/master@{#364918}