Force Account Consistency off for Logout call on iOS.
The GaiaAuthFetcher Logout call needs Account Consistency to be
disabled to work properly. On iOS, the state of Account Consistency is
handled by cookies set on the various Google domains. Sadly, the cookie
clearing request is not synchronized with the Logout call and there is
a race between the two.

This CL forces the cookie to be invalid during the Logout request to
ensure the accounts will be cleared. The cookie will eventually be
invalidated in the WKWebView cookie jar when AccountConsistencyService
finishes.

BUG=569501

Review URL: https://codereview.chromium.org/1522873002

Cr-Commit-Position: refs/heads/master@{#365248}