Android Chromoting: Request GET_ACCOUNTS permission if needed.
Now that the app is targeted at API level 23, it needs to request any
permissions that it needs at runtime and cannot assume that all
permissions are granted. This CL requests the GET_ACCOUNTS permission
in AccountSwitcherBasic, if it isn't already granted.

The offical Google account-switcher doesn't seem to require this
permission. So this CL only updates the basic account-switcher with
minimal changes needed for correct functioning. No new translatable
strings are added, for example, which would normally be used to
explain to the user why the permissions are required.

BUG=546378

Review URL: https://codereview.chromium.org/1528783002

Cr-Commit-Position: refs/heads/master@{#365595}