[Android] Add record functionality to logcat monitor.
Because Android Chromium infra may be switching over to swarming,
we will need logic to record logcats that is triggered by
individual scripts and not the bots. This is because in the future
the bot that trigger the tests may not have the devices attached to
them.

BUG=

Review URL: https://codereview.chromium.org/1484253003

Cr-Commit-Position: refs/heads/master@{#365878}