Add support for (de)serializing LayerSelection.
As part of serializing cc::LayerTreeHost, we also need to
serialize the Selection<LayerSelectionBound>, which is
typedef'ed to LayerSelection.

We only support this type of the templated Selection for now,
so the code lives together with the LayerSelectionBound.

BUG=561210
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1531413002

Cr-Commit-Position: refs/heads/master@{#366202}