Autofill's CountryCode should use scoped_ptr for owned pointers
This CL turns a map of raw pointers to a map of scoped_ptrs. It is a clean-up
which is unlikely to fix the associated crash, but the explicit semantics
might help discover the real cause. And besides, using scopers for owned
pointers is a good practice.

BUG=571371

Review URL: https://codereview.chromium.org/1541843002

Cr-Commit-Position: refs/heads/master@{#366414}