Use content::NavigateIframeToURL in ProcessManagerBrowserTest
In 1413853005 I introduced NavigateIframeToURLAndWait in a test because
the current content::NavigateIframeToURL implementation does not account
for nested frames. https://codereview.chromium.org/1522443004/ fixes the
issue, so let's use it.

ChromeSitePerProcessTest.PluginWithRemoteTopFrame has been updated
because I noticed that the flash object that was supposed to be loaded
did not load because the file path was incorrect.

Review URL: https://codereview.chromium.org/1502963004

Cr-Commit-Position: refs/heads/master@{#366416}