MediaRecorder: new tests for illegal function invocation
This CL introduces new tests for MediaRecorder regarding illegal
use of start() and stop() functions
when the MediaRecorder is not in a proper state to execute
the said functions.
Such illegal use of those functions throw a DOM Error.
Basically we must not start() when already recording,
we must not start() when paused and we must not stop()
when MediaRecorder is inactive.

BUG=548914
TEST=On Linux host and on trybots.

Review URL: https://codereview.chromium.org/1531263007

Cr-Commit-Position: refs/heads/master@{#366431}