Fix computation of min|max-content contribution of non-replaced blocks
Currently Blink always returns the min preferred logical width for the
min-content contribution (and the max preferred logical width for the
max-content contribution) for non-replaced blocks. That is not correct
according to specs https://drafts.csswg.org/css-sizing/#block-intrinsic.

The min-content and max-content contributions actually depend on the
computed inline size of the block:

* for min-content,max-content or definite sizes: min-content and max-content
contributions are that size plus border, margin and padding.
* otherwise: min-content contribution is the min-content size and
max-content contribution is the max-content size (in both cases plus border, padding
and margin).

Review URL: https://codereview.chromium.org/1504403002

Cr-Commit-Position: refs/heads/master@{#366753}