Ignore aborts with an abort timestamp before navigation start
As part of crrev.com/1543833002, it was discovered that in some cases
we may record a navigation abort time before navigation start.

In WebNavigationApiTest.CrossProcessAbort, pending navigation A is
initiated, then pending navigation B is initiated, then A commits.
This causes B to abort. When attributing abort time and cause for B,
we use A's navigation start time. However, since A started before B,
this means that the attributed abort time is before the nav start time,
which doesn't make sense. This change addresses this case by ignoring
such cases, and incrementing an error counter so we can track how
frequently this happens.

BUG=571647

Review URL: https://codereview.chromium.org/1542033003

Cr-Commit-Position: refs/heads/master@{#366822}