Change the equation for computing sigma from blur radius
In a previous CL: https://codereview.chromium.org/1502283002/, this equation
was changed and now when the blur radius is 2px the blur effect completely
disappears. This CL change the equation back to the original one but add
a special case handler so that when the blur radius is 0 there is no blur
effect at all.

Also, together with this change, a bunch of layout test will need to
rebaseline.

BUG=572822

Review URL: https://codereview.chromium.org/1556523002

Cr-Commit-Position: refs/heads/master@{#367188}