Catch null return value from computeComputedStyle.
This patch is in response to a Crash Report, so I don't have a repro. I
haven't been able to trigger the condition locally. However it seems
that other call sites to ensureComputedStyle (which computeComputedStyle
calls through to) perform a null check, so it's sensible that we need to
here as well.

Unfortunately, no repro means no test case :(

BUG=571040

Review URL: https://codereview.chromium.org/1547183002

Cr-Commit-Position: refs/heads/master@{#367288}