Vp9Parser: correct reading of `log2_tile_rows` field
Previously the parsing of the `log2_tile_rows` field in the
uncompressed header would consume one bit too many in the case
where log2_tile_rows > 0. This corrupts the reading of the rest of
the header and causes playback to fail. Here the reading logic is
corrected.

BUG=569149
TEST=play the linked media in the bug on samus

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>

Review URL: https://codereview.chromium.org/1516413003

Cr-Commit-Position: refs/heads/master@{#367404}