Optimizations to Spellchecking Chunk Generation
1. The existing implementation frequently does translations between
offsets and positions, and each translation needs to scan from the
beginning of the text paragraph. The patch avoids such translations
completely.

2. Only the first chunk needs to have both of its start and end
positions expanded to sentence boundaries. For each subsequent chunk, we
only need to expand its end position to sentence boundary. Hence, we can
reduce the number of calls of startOfSentence to 1 regardless of the
number of chunks generated.

BUG=109587

Review URL: https://codereview.chromium.org/1563783003

Cr-Commit-Position: refs/heads/master@{#368035}