Fixes bug where client area would not get set correctly
The client area is updated when the bounds change, further to
correctly calculate the client area the NonClientView has to have been
created. Prior to this fix if the window was sized during
InitNativeWidget() and never changed after that, then the client area
was not correctly set because during InitNativeWidget() the
NonClientView has not been created.

The fix is to add a NativeWidgetPrivate function that Widget calls
after it has created the NonClientView.

BUG=none
TEST=none
R=erg@chromium.org

Review URL: https://codereview.chromium.org/1565243002

Cr-Commit-Position: refs/heads/master@{#368141}