Enable warning on reserved user-defined literals everywhere but CrOS.
This was disabled because a dbus header was missing a few spaces.
The version of the dbus header in the sysroot has them though,
and we use the sysroot everywhere except in Chrome OS-on-Linux
builds these days.

The spaces were added here, almost 3 years ago by now:
http://cgit.freedesktop.org/dbus/dbus/commit/?h=dbus-1.4&id=51b88b4c7919487290c0862b013cd8e7cd2de34b

No behavior change.

BUG=263960, 573778

Review URL: https://codereview.chromium.org/1570193002

Cr-Commit-Position: refs/heads/master@{#368408}