Revert 3 CLs for shouldOverrideUrlLoading implementation.
Since https://codereview.chromium.org/1424263003 does not handle
overriding about:blank loads (or rather, about:blank loads are
handled synchronously and can thus be finished before we have a chance
to override them) we cannot land that patch without potentially breaking
lots of apps using WebView.

The second CL being reverted is a fix for the first one to post
didStopLoad and didFailLoad in the correct order.

The third CL being reverted here is a small change removing code that
is not used by https://codereview.chromium.org/1424263003 but is used
now that we revert that CL.

This reverts commit bae3ff13e6959b713b872df3c4a734d6a1ea01ad.
This reverts commit e1da704f67f34022b7aaf3afe277c86217844114.
This reverts commit c002c7ea9c5a3dbeb665e4070d852e6724d1ff86.

TBR=palmer@chromium.org,jam@chromium.org
BUG=325351,462306

Review URL: https://codereview.chromium.org/1530183004

Cr-Commit-Position: refs/heads/master@{#368849}