aura: Propagate PointerEventType through the MotionEvent
When the TouchEvents are aggregated into a MotionEvent the pointer type
information was being lost. This change maps the PointerEventType on the
PointerDetails for the TouchEvent to the MotionEvent::ToolType.

A follow-up CL @ https://codereview.chromium.org/1575993002 enhances the
MotionEventAuraTest.ToolType test when the ability to override the
PointerDetails on a TouchEvent is added.

TEST=Interactive testing on link with http://rbyers.github.io/eventTest.html
and event_unittests:MotionEventAuraTest.ToolType
BUG=575165

Review URL: https://codereview.chromium.org/1568843002

Cr-Commit-Position: refs/heads/master@{#369128}