MediaRecorder: fire onError if the amount of tracks of the MediaStream varies
MediaRecorder is supposed to fire an Error event if the amount
of tracks of the associated MediaStream varies. This CL adds
monitoring of said amount and Error firing.

There's no need to stop the MediaRecorder: adding a Track
would not start automatically recording it, and removing it
guarantees it would not be recorded not recording. Moreover,
the spec does not say if MRecorder should stop if the
#Tracks varies, and also instructs to throw an Exception,
which is not possible since the error is asynchronous.

Also this CL removes superfluous test variables.

BUG=545156

Review URL: https://codereview.chromium.org/1578973002

Cr-Commit-Position: refs/heads/master@{#369333}