Add support for (de)serializing LayerTreeHost.
As part of the cc commit flow, we need to be able to serialize
the LayerTreeHost. Not all state is necessary to serialize to
be able to do a commit on the client side, so only some members
are included in the proto.

BUG=561210
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1519293002

Cr-Commit-Position: refs/heads/master@{#369450}