Report VP8 and VP9 capabilities for non-accelerated decoders in AVDA
We need them because we use AVDA for EME even for non-acelerated
decoders. This CL reports VP8 and VP9 capabilities for all decoders,
but succeeds AVDA VP8/9 initialization only for the accelerated
ones (as before) or for both accelerated and non-accelerated if the
stream is encrypted.

BUG=577424

Review URL: https://codereview.chromium.org/1583083002

Cr-Commit-Position: refs/heads/master@{#369525}