Fix UAF in MimeHandlerViewContainer
The following html snippet is sufficient to reproduce, given a valid pdf:
<body onload="i.remove()"><iframe id=i src="your-valid-url.pdf">

MimeHandlerViewContainer creates an AssociatedURLLoader to load content when the plugin is in an iframe. When MimeHandleViewContainer::didFinishLoading() is called, it clears the AssociatedURLLoader, which instructs blink to cancel the load. Blink then checks whether we can now run the load event (in FrameLoader::checkCompleted). The onload event has already run in the iframe (just before didFinishLoading was called), so we give the parent a chance to run its onload. It does, which detaches the iframe, deleting the plugin hierarchy and MimeHandleViewContainer. The stack unwinds to MimeHandleViewContainer::didFinishLoading(), which UAFs.

There are two ways to fix this: The obvious option is to make MimeHandleViewContainer more defensive. Alternatively, we could view this as a bug in blink: the AssociatedURLLoader doesn't block the load event, so why does it lead to a FrameLoader::checkCompleted() call? This patch implements the latter, because that is clearly a bug and is required for this UAF.

BUG=573886

Review URL: https://codereview.chromium.org/1582693002

Cr-Commit-Position: refs/heads/master@{#369915}