ui: Fix TouchEvent PointerDetails creation from NativeEvent
This change replaces the GetTouchRadiusX/GetTouchRadiusY/GetTouchForce platform
functions with a single GetTouchPointerDtailsFromNative function, which aligns
with the behaviour for MouseEvent.

The default implementation for this (used on ozone) will correctly preserve the
pointer type for the TouchEvent now.

This change also modifies the default radius values for Mac OS X and the stub
implementation to match the documented default of 1.0.

BUG=575162
TEST=events_unittest passed on all platforms, interactive testing of Pointer
Events of touch events on link.

Review URL: https://codereview.chromium.org/1586653002

Cr-Commit-Position: refs/heads/master@{#370110}