Don't change layout size due to top control show/hide.
This patch changes Chrome's top control (URL bar) behavior on Android
to match Safari's in how it affects layout size. To summarize, the
layout size or, equivalently, the initial containing block (ICB), is no
longer dynamically adjusted when the top controls are shown/hidden.
Instead the ICB is statically sized to be the size of the viewport when
top controls are showing. position: fixed elements are an exception,
they continue to work as they do today, resizing to mach the full
viewport size in response to top controls.

Additionally, vh units were changed in a similar way; they are now also
statically sized but to the size of the viewport with top controls
hidden.

See https://github.com/bokand/URLBarSizing for full details.

Intent to ship:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/BK0oHURgmJ4

BUG=428132

Review URL: https://codereview.chromium.org/1573283002

Cr-Commit-Position: refs/heads/master@{#370425}