Mac: Fix IOSurfaceIsInUse flakes
Recall that the timeline for a cc::Resources being used as an overlay is
as follows
1. Resource is sent from the Renderer to the Browser.
2. Resource is locked in GLRenderer::pending_overlay_resources_ while
   preparing to swap.
3. Resource's lock is moved to GLRenderer::swapped_overlay_resources_
   when GLRenderer::SwapBuffers is called.
X. In the GPU process, the Resources' IOSurface is set to a CALayer's
   contents in GLSurface::SwapBuffers, making IOSurfaceIsInUse return
   True until the WindowServer is done with Resource.
4. Resource's lock is released in GLRenderer::SwapBuffersComplete when
   the GPU process has completed its call to GLSurface::SwapBuffers
5. Resource is returned to the Renderer process
6. In the Renderer process, the resource may be re-used as soon as it is
   returned, as soon GpuMemoryBuffer::IsInUseByMacOSWindowServer
   indicates that the resource is not in use. This corresponds to
   checking IOSurfaceIsInUse

That's great, except that in X, we don't actually set the IOSurface to
be the CALayer's contents in GLRenderer::SwapBuffers. Rather, we put the
IOSurface in a CALayerTree:ContentLayer, and, at some point in the
future (about 1 vsync later), we actually set the IOSurface to be the
contents of a CALayer.

So, if it takes less than 1 vsync between returning the resource in
step 5 and re-using it in step 6 (which is basically always true), then
we'll end up re-using an IOSurface that still has yet to be sent to the
WindowServer (flashing, sadness, etc, follow).

So the fix is pretty simple. At X, make CALayerTree::ContentLayer
increment the IsInUse count. And keep it incremented until the
CALayerTree::ContentLayer is destroyed. In principle, we could just
increment IsInUse until the IOSurface is set to a CALayer's contents,
but this is easier to verify as correct.

Do the increment in all constructors, do a decrement in the destructor.
Make the CALayerTree::ContentLayer::io_surface be const, so that we
can be sure that it is the same IOSurface that we are accessing
throughout.

BUG=577121

Review URL: https://codereview.chromium.org/1608293002

Cr-Commit-Position: refs/heads/master@{#370623}