ui/base/models: rewrite TreeNode without using ScopedVector
Now that we are using C++11, moveable types inside containers are supported
and ScopedVector will be removed. So we should migrate to a solution
that does not use ScopedVector.

Unfortunately children() accessor complicates the things and does not
allow us to easily use std::vector<scoped_ptr<NodeType>>, so another
solution with STLDeleteElements was choosen.

BUG=554289
TEST=ui_base_unittests --gtest_filter=*Tree*
R=sky@chromium.org

Review URL: https://codereview.chromium.org/1619733003

Cr-Commit-Position: refs/heads/master@{#370769}