[Chromecast] Break ozone -> chromecast/ dependency
This is to help break the ozone -> chromecast -> media -> ozone
cycle.  A bunch of offline discussion led to agreement that
the media -> ozone dependency is undesirable, and could be removed
by refactoring some of the graphics types that media requires.
But that's a lot of work, and the ozone -> chromecast part of the
cycle isn't great either.

Essentially, we want Cast ozone to be able to notify interested
parties of overlay geometry as it is composited.  The video plane
can be injected as such.  It uses globals because there is no control
over Ozone platform constructor signature.

BUG=

Review URL: https://codereview.chromium.org/1583553003

Cr-Commit-Position: refs/heads/master@{#370873}