Delay posting errors from AVDA.
For full screen transitions and other cases where we will lose the
SurfaceView, we don't want the pipeline to transition to the error
state if it's about to reset us anyway.  This CL defers error
reporting by two seconds to give the rest of the pipeline a chance
to deal with the transition.

Ideally, we would not depend on the pipeline to know that the AVDA
(probably) entered an error state and needs to be reset.  However,
for our initial SurfaceView implementation, it does.  AVDA should
signal that it requires a restart, and probably defer errors only
in that case.

The GpuVideoDecoder also now dispatches a bitstream buffer upon
receipt of an error from the VDA.  Otherwise, if the number of
outstanding bitstream buffers is at its limit, then no further
bitstream buffers will be sent in until the VDA finishes one.
Since that's the only way to fi../../third_party/WebKit/Source/core/css/CSSFontSelector.cpp:27:nd out about a decode error, and
since there is no VDA after an error occurs, GVD now returns one
immediately with the error.

BUG=575243, 574253

Review URL: https://codereview.chromium.org/1570613002

Cr-Commit-Position: refs/heads/master@{#371272}