[Media Router] Handle route creation where resolved route is not for display.
This change handles route creation where the initially resolved route is not marked for display. This keeps track of the route id and checks if the route is updated to be ready for display.

In the case where the route is never ready for display, route creation times out and fails. The MRP will send an issue to the Media Router if this happens.

Updated reporting for relevant UMA metric.

BUG=579138

Review URL: https://codereview.chromium.org/1605133002

Cr-Commit-Position: refs/heads/master@{#371385}