If you reload an offline page while online, load the online version.
If you have an offline page showing, but connectivity has returned, when
you pull down on the page to refresh, we should reload the online version.

This change catches the onPageLoadStarted event to check for that case.
The onLoadUrl event does not get called on a reload, and the onUpdateUrl
event is also not called on a reload, so onPageLoadStarted is likely
our best candidate.

BUG=565529

Review URL: https://codereview.chromium.org/1586783003

Cr-Commit-Position: refs/heads/master@{#371572}