[MD] Don't use new OTR native-theming when using a custom browser theme.
The location bar, omnibox, and find in page bar should look the same in
incognito and normal windows if we're using a custom theme. Since
changing to or from a custom theme can now affect the window's decision
of which NativeTheme to use, we have to make GetNativeTheme()
overrideable as well as propagate NativeTheme changes when the browser
theme changes.

BUG=581559
depends on http://crrev.com/1638063003#ps1

Review URL: https://codereview.chromium.org/1637203002

Cr-Commit-Position: refs/heads/master@{#371990}