Fix scroll restoration when exiting fullscreen mode.
There's two problems in Chrome/Blink that cause the existing scroll restoration
not to work on desktop platforms:

In Blink, when entering fullscreen mode, FullscreenController saves the current
scroll offset so that it can restore it later. However, this was done in
FullscreenController::didEnterFullScreen, which is called only after WebViewImpl
is resized, meaning that the scroll offset may already be wiped out.

In Chrome, we explicitly send a WasResized message to ensure that the renderer
is notified of the fullscreen change in cases where toggling fullscreen doesn't
actually change the renderer's size (e.g. going from tab fullscreen to browser
fullscreen). However, this means that the renderer gets two separate resize
events for the fullscreen toggle: one that toggles fullscreen and another for
the size change. Thus, Blink tries to restore the scroll offset before the
renderer is resized which might then clamp that scroll offset.

For the first issue, I've moved the scroll offset saving logic into
FullscreenController::enterFullScreenForElement which is called before sending
the request IPC to the browser.

For the second issue, I added a will_cause_resize parameter to
WebContents::ExitFullscreen that tells the method whether exiting fullscreen will
potentially cause a window resize. Based on this, WebContents can choose to
explicitly send the WasResized only in cases where it will not occur otherwise.
This ensures the Resize message includes both the fullscreen state change and
size change in one message.

BUG=142427
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review URL: https://codereview.chromium.org/1488653002

Cr-Commit-Position: refs/heads/master@{#372146}