Handle subresource cert errors even when there is no ResourceLoader
Previously, MixedContentChecker relied on a ResourceLoader to notify the
embedder about subresources laoded with certificate errors. When a
ResourceLoader wasn't present -- notably, for resources served from the
memory cache -- the embedder wasn't getting notified at all.

MixedContentChecker was using the ResourceLoader only to figure out the
request's |frameType| and |requestContext|. This CL makes it so that
FrameFetchContext passes in those values directly instead of passing in
the ResourceLoader.

BUG=477868

Review URL: https://codereview.chromium.org/1648613003

Cr-Commit-Position: refs/heads/master@{#372349}