cc: Fix a use-after-free in discardable image map.
In display item list world, we may use a temporary paint during a save
which is gone by the time we process the paint, causing use after free.

This patch ensures that we take a copy of the paint, which means that
when we process the paint, we won't use after free.

R=enne, wkorman
BUG=582606
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1651533003

Cr-Commit-Position: refs/heads/master@{#372769}