Force multi-lingual support for spell-checking
The current spellchecker behavior (on certain platforms) can switch between
single-language spellchecking and multi-lingual spellchecking based on a
#flag. There is also a checkbox to disable spellchecking globally but
only in single-language mode.

This CL removes the option to have single-language spellchecking (it forces
multi-lingual) but, of course, the user is free to have only a single
language listed in their list, and/or to enable it/them.

However, this CL preserves the ability to disable spellchecking globally, by
making the checkbox permanently visible.

BUG=579346

Review URL: https://codereview.chromium.org/1606243003

Cr-Commit-Position: refs/heads/master@{#372803}