Report uncompressed header bytes in OnBlockHeaderEnd().
Modify HpackDecoder to report uncompressed header bytes when used with a
SpdyHeadersHandlerInterface.  This makes HpackDecoder more closely aligned with
SpdyHeadersBlockParser, which always reports the size of the uncompressed (but
encoded) header block. There's no way to separate the compression from the
encoding in HPACK, so this is the best we can do.

This CL lands server change 113275350 by birenroy.

BUG=488484

Review URL: https://codereview.chromium.org/1656253004

Cr-Commit-Position: refs/heads/master@{#373087}