Remove unused html format support from device_event_log
Some code cleanup that will allow us to move device_event_log out of
src/components if we need to. It turns out we switched to passing
events to the UI as JSON but never removed the HTML format support
which requires a src/net dependency.

BUG=none

Review URL: https://codereview.chromium.org/1663073004

Cr-Commit-Position: refs/heads/master@{#373871}