Communicate whether passive event listeners exist to cc.
cc does nothing with these new fields being set other than plumbing them
through the pipeline to be later queried on the layer_impl.

Add an additional field whether touch event listeners exist on the layer
(which unfortunately is effectively implement as an entire document field
in the blink EventListenerRegistry; this can be improved). There still
exists TouchRegionRects that indicate the areas of blocking event
listeners; but whether a passive event listener exists will be indicated
in this field. It is intended that this model will be replicated
with property trees when it is ready.

BUG=489802
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1577263004

Cr-Commit-Position: refs/heads/master@{#373915}