SafeBrowsing: Remove unused histogram bucket.
MIN_WAIT_DURATION_ERROR was added in cl/1563763001 to account for v4 request
errors. V4 requests have since been moved to their own enum in cl/1657233002
and MIN_WAIT_DURATION_ERROR is unused in the SBGetHashResultType enum. This
bucket never recorded any metrics because the code is not live yet, so it is
safe to change the enum.

BUG=561867,543161

Review URL: https://codereview.chromium.org/1664003002

Cr-Commit-Position: refs/heads/master@{#374082}