Code clean up: Move Property -> InterpolationTypes mapping out of StringKeyframe
This patch has no change in behaviour.
This change moves the mapping from PropertyHandle to list of applicable
InterpolationTypes to its own file instead of residing in StringKeyframe.

StringKeyframe is the only consumer of this mapping but its purpose is
distinct enough to warrant its own class.

Review URL: https://codereview.chromium.org/1622183003

Cr-Commit-Position: refs/heads/master@{#374310}