Indexed DB: Remove transaction throttle
A transaction throttle was introduced in https://crrev.com/1511973004
to mitigate behavior where a renderer was issuing transactions faster
than they could be processed, leading to eventual OOM. Unfortunately,
it looks like some apps are relying on behavior that trips this
throttle, so we need to rethink it. Remove the throttle, but keep
recording UMA stats.

R=cmumford@chromium.org
BUG=338800,584768

Review URL: https://codereview.chromium.org/1671193003

Cr-Commit-Position: refs/heads/master@{#374443}