Add path open errors from the permission broker to the device log.
This change adds an additional ErrorCallback parameter to the permission
broker client so that the caller can get a better log message when its
request to open a device node fails.

This will hopefully help to determine the cause of the many "Did not get
valid device handle from permission broker" errors that some users are
seeing.

BUG=570784

Review URL: https://codereview.chromium.org/1681383002

Cr-Commit-Position: refs/heads/master@{#374571}