ChromeVox Next: apply aural styling to some output rules.
This cl in-lines aural styles directly into the speech format rules. The speech operators themselves apply to the rest of the string (similar to some embedded tts commands on Mac).

We style the non-primary text (non-name) speech with some pitch changes. The only exception are headings. We shift downward as heading levels go down and apply that change to the name.

Also, we directly apply some tts attributes when we see the $role or @<message_id> tokens. This is to save us from adding them to every rule.

Note the use of the ! prefix. It adds the following token directly as a tts property.

Valid properties are scattered inside of ChromeVox source and also encompass tts api properties.

Currently, the !<tts_property> takes one or no arguments. We attempt to parse the arg as a float (and if that fails) we use it as an attribute lookup. The result of that lookup will be parsed as a float.

TEST=observe styles while browsing and while navigating in views.

Review URL: https://codereview.chromium.org/1648103002

Cr-Commit-Position: refs/heads/master@{#375699}