Add chrome://physical-web as a diagnostics endpoint
The diagnostics page checks whether prerequisites for the Physical Web
feature are met and reports this information back to the user.
If UrlManager has cached URL data, the cached data is also reported. If
there are URLs that have been found but cannot be included in the URLs
list, the URL will be reported along with the reason for its exclusion.

The chrome://physical-web endpoint is only accessible when the Physical
Web chrome flag is enabled (chrome://flags/#enable-physical-web)

BUG=577065

Review URL: https://codereview.chromium.org/1591143002

Cr-Commit-Position: refs/heads/master@{#375705}