tools/android/loading Ignore timings for requests coming from cache
The requests coming from Blink cache have stale timings: these
are the timings corresponding to the request that put the content
in the cache, rather than of the request that got them from the cache.

Ignoring these timings allows to fallback to the timestamp, which is
more accurate (although not perfect), and fixes issues where the
inconsistent timings was breaking initiators.

Review URL: https://codereview.chromium.org/1708233002

Cr-Commit-Position: refs/heads/master@{#376187}