Assign correct RenderWidget pointer during RFImpl to RFProxy swap
Currently we observe a crash when a local frame with a remote parent is swapped
to a remote frame, and then back to a local frame. This is because the
RenderFrameProxy receives the previous RenderFrameImpl's RenderWidget pointer, which
can become invalid when the RenderFrameImpl is destroyed.

This CL assigns the correct RenderWidget pointer to the RenderFrameProxy
at creation time.

BUG=585654

Review URL: https://codereview.chromium.org/1702143002

Cr-Commit-Position: refs/heads/master@{#376218}