[Layout API] Add api/LayoutBoxModel.h and use in FrameView::invalidateViewportConstrainedObjects
This adds LayoutBoxModel as a descendant of LayoutItem in the new
layout API and provides a single method, layer(), which is used by
FrameView. It also adds a few methods to LayoutItem to support
descendants (protected methods to get the underlying layout object,
a public isBoxModel method).

There are no functional changes.

BUG=499321

Review URL: https://codereview.chromium.org/1712703003

Cr-Commit-Position: refs/heads/master@{#376572}