Fix javac sometimes not create all outputs with enable_incremental_javac
The problem is that sometimes there are .apks that don't have any
resources (e.g. test apks), and so there is a javac.py call with a
single .srcjar, which has no sources in it. Ninja is expecting the call
to create a .jar.pdb file, so just create an empty file to appease it.

Without this, running "ninja chrome_public_test_apk" multiple times
causes ninja to do the same work over and over again, rather than saying
there's no work to do

BUG=none

Review URL: https://codereview.chromium.org/1706363003

Cr-Commit-Position: refs/heads/master@{#376984}