Make two fixes to ChromeVox Panel.
1. ChromeVox+. in a brand new session.

Observe that the global.backgroundObj.savedRange_ is set to the hosting webView to the Panel. Further logging should show that we get a focus event on the webView. As a result, when you press escape, you actually stay inside of the panel.

Note that subsequent invokations of ChromeVox+. does not result in this focus event.

2. ChromeVox+. for the first time. Notice that focus does not land on the first menu item. We don't get a focus event for it (just a load complete on the root). Make it so that the load complete event checks focus via getFocus().

Review URL: https://codereview.chromium.org/1719143004

Cr-Commit-Position: refs/heads/master@{#377135}