Avoid marking LayoutFullScreen as orthogonal writing-mode roots
This patch avoids marking LayoutFullScreen and
LayoutFullScreenPlaceholder as orthogonal writing-mode roots.

These objects are set to horizontal flow and inserted in the middle of
the layout tree. If the insertion point is in vertical flow, this
process creates two orthogonal writing-mode roots.

This creates inefficient layout, and sometimes FrameView fails to track
such orthogonal writing-mode roots due to the way it modifies the tree.

Since the fix is similar to crrev.com/376328, two new methods are added
to not to create unnecessary orthogonal writing-mode roots:
* LayoutObject::addChildWithWritingModeOfParent()
* LayoutObject::setStyleWithWritingModeOfParent()

The fix for LayoutScrollbarPart in crrev.com/376328 is aligned to the
same patten in this patch.

BUG=588552

Review URL: https://codereview.chromium.org/1721653002

Cr-Commit-Position: refs/heads/master@{#377385}