Update layout_test TestExpectations and webkitpy support for new Mac OSes.
With this CL we drop support for running the layout_tests on Mac10.6, 10.7,
or 10.8, and update the (testing) code to support 10.11. The "Retina" port
is also now expected to be running 10.11.

This CL contains a new Mac10_11_Expectations file that contains all of
the failures I expect to see on a 10.11 bot (Retina or non-Retina),
based off of a local (non-Retina) build. I will update the file once
the 10.11 builders have cycled and we have more accurate results.

I expect that at this point the 10.11 bots will remain largely useless.
None of the 10.10-based baselines in platform/mac have been moved to
platform/mac-mac10.10, and none of the failures will have been triaged
or rebaselined. That said, at least it's a start and will allow us to
make progress.

I'm marking this NOTRY=true because the existing 10.6/10.7/10.8 bots
will break and we need to restart the chromium.webkit bots anyway.

R=ojan@chromium.org
TBR=ericwilligers@chromium.org, ojan@chromium.org
BUG=516463, 589709
NOTRY=true

Review URL: https://codereview.chromium.org/1733163002 .

Cr-Commit-Position: refs/heads/master@{#377494}