Added a deprecation warning label to fullscreen/mouselock settings.
Currently, on Windows, Linux and Chrome OS, these settings are still
present, but not used (so that users can view and edit the exceptions
stored in their profile). Adding a warning to explain that they are
unused, on those platforms. Also moved the sections to the bottom and
renamed "Mouse cursor" to "Mouse cursor lock" for clarity.

BUG=589710

Review URL: https://codereview.chromium.org/1740573003

Cr-Commit-Position: refs/heads/master@{#377788}