Enable the new-app-list-mixer flag on Chrome OS.
This changes the way the Launcher ranks and mixes search results. For
details on the change, see https://crrev.com/329601.

It will still respect both the flag and the field trial, but as the
trial is not currently active, the default should be respected for all
users who haven't explicitly set the flag.

BUG=487494,460451,422610,338418

Review URL: https://codereview.chromium.org/1744493002

Cr-Commit-Position: refs/heads/master@{#377833}