Adding filtering behavior for NTP suggestions.
For child accounts, do not suggest popular sites. For Whitelist entry
point suggestions filter the web sites that were already suggested. For
most visited suggestions, filter suggestions that are blocked for the
child (for example if a child visited a web site that would appear on
most visited, but the parent manually blocks it).

BUG=586097

Review URL: https://codereview.chromium.org/1731963002

Cr-Commit-Position: refs/heads/master@{#378016}