Only commit sync preference settings in onStop()
We only want to commit the user's sync preferences when the user is finished
setting up sync. With the UX changes to the sync setting screen, we would
inadvertently start syncing if the user taps on the preference to switch
sync accounts.

Also fixing a crash where if we entered SyncCustomizationFragment while sync
was off, we would not update the sync preference's list of possible accounts
to sync to.

BUG=588971, 557784

Review URL: https://codereview.chromium.org/1728493002

Cr-Commit-Position: refs/heads/master@{#378197}