mojo: Sketch a profile application.
This application is bound to an individual profile data dir. While it
isn't yet sandboxed to that location, that's our long term intention
with this design. This application colocates the Profile service and the
LevelDB service so that these will be in the same process.

The way we register a Browser Context to a user id works for now, but
will need to be redone once we start spawning processes for the Profile.

BUG=585587

Review URL: https://codereview.chromium.org/1741953002

Cr-Commit-Position: refs/heads/master@{#378620}