Extend FontCache fallback API to support FontFallbackPriority
This is take two and now compatible with Android. Instead of retrieving
a list of font family candidates I am now extending the existing
character fallback API with the FontFallbackPriority argument. Depending
on the platform, this gives us better search options for finding the
fallback font. On Android specifically, we can search for the character
plus a specific locale for retrieving the color emoji font without an
explicit family name. In Windows, we can better integrate with the
existing fallback logic.

Unfortunately, this is change is hard to test on the bots as an Emoji
font is not available on all Linux bots and blink_platform_unittests do
currently not execute on Android. But the change was manually verified
on Mac, Windows, Linux and Android.

BUG=580536

Review URL: https://codereview.chromium.org/1757703003

Cr-Commit-Position: refs/heads/master@{#378826}