Fix SVGDocument return title that is not a child of the root element
The title element should be a child of the root element.
But currently return invalid title that is not a child of the root element.

Because on getting, if the root element is an svg element in the SVG
namespace, then let value ba a concatenation of the data of all
the child Text nodes of the first title element in the SVG namespace
that is a child of the root element.[1]

And on setting, let element be the first title element in the SVG
namespace that is a child of the root element, if any.
If there isn't one, create a title element in the SVG namespace,
insert it as the first child of the root element, and let element
be that element.[1]

[1] https://html.spec.whatwg.org/multipage/dom.html#document.title

BUG=543061

Review URL: https://codereview.chromium.org/1752713002

Cr-Commit-Position: refs/heads/master@{#379301}