Improve currentTime() accuracy with large audio buffer sizes.
On platforms which use a large audio output buffer size, the time
between Render() callbacks can cause a significant lag in updates
to the currentTime() value. We should instead return a value which
is compensated for the elapsed real time since the last Render()
callback.

In days of yore we didn't want to do this since the time could go
backwards, which can cause problems for Blink and other consumers
of the currentTime() value. The solution here is to add clamping
and estimate current time based on the elapse of real time.

BUG=263654
TEST=new unittest.

Review URL: https://codereview.chromium.org/1763843002

Cr-Commit-Position: refs/heads/master@{#379345}