[css-grid] Fix auto-track sizing with min-size:auto and specific sizes
Specs recently changed the way auto tracks are sized. In the previous
versions only the min-width|height of the items spanning through the auto tracks
were used to size them. The new text specifies that for items with
a specified minimum size of auto, the behavior is equivalent to
a min-content minimum.

This means that from now on, auto tracks with min-size:auto will no longer
be smaller than min-content tracks (which was pretty weird from the user
POV).

BUG=544539

Review URL: https://codereview.chromium.org/1609153003

Cr-Commit-Position: refs/heads/master@{#379554}