[Extensions UI Views] Generic-ify a toolbar actions bar bubble
There's a generic ToolbarActionsBarBubbleDelegate, but the view for the bubble
was linked to the ExtensionIconSurfacingBubble. There's no reason it has to be,
and this can be re-used for other bubbles that dangle off the toolbar. Make the
view more abstract (we already do this on mac), and add a few more tests for it.
Next step will be to use this in click-to-script work.

BUG=460306

Review URL: https://codereview.chromium.org/1771033002

Cr-Commit-Position: refs/heads/master@{#379936}