Remove static int views::BubbleBorder::GetCornerRadius()
It returns 4, but the corner radius isn't constant.

Only ExtensionPopup uses it; with the default shadow type of
SMALL_SHADOW, which actually has a corner radius of 2 anyway.

ExtensionPopup could use GetBubbleFrameView()->bubble_border()->
GetBorderCornerRadius() to get the actual corner radius, but that's not
available until after the frame view is created, and ExtensionPopup uses
GetCornerRadius() to derive the margin that is passed into the
BubbleFrameView constructor (i.e. before it's created), so it can't.

The default margin just comes from views' layout_constants.h, so using a
constant for the ExtensionPopup makes sense too. Let's do that.

BUG=None

Review URL: https://codereview.chromium.org/1761203002

Cr-Commit-Position: refs/heads/master@{#380077}