Make sure cancel event fires in layout test.
The web-animations-api/player-cancel-event layout test can trivially
pass if the animation oncancel event is never called. This patch adds
a bool flag to the test to additionally test that the oncancel event
has been called, as well as testing that the event's properties are
correct.

BUG=

Review URL: https://codereview.chromium.org/1765383005

Cr-Commit-Position: refs/heads/master@{#380082}