The command line flag is now set during the test startup.
These two tests replaced the global command line object, which might
cause heap-use-after-free of the temporary object at the end of the
test. For instance, ChromeExtensionsClient::GetWebstoreUpdateURL()
tries to read the command line flags, and does so asynchronously, not
in the main thread of the test.

The proposed patch eliminates use of the ScopedExperimentalCommandLine
at the test scope level, thus minimizing the risk of improper access to
memory.

BUG=
R=finnur@chromium.org

Review URL: https://codereview.chromium.org/1752293002

Cr-Commit-Position: refs/heads/master@{#380096}