Temporarily allow loading v2 apps in tabs.
Add a UMA to understand how widely this is being used.
The window.open() call from a background page takes a different
code path than the non-background page one. So collect metrics
for both code paths.

The reason for this temporary relaxing is:  apps can use this to
use cast API. The alternative to this is to use Presentation API,
which requires Media Router to go stable. Since MR is not
available on stable yet, we're going to temporarily allow loading
v2 apps in tabs.

This CL will be reverted once MR goes stable.

TBR=asargent@chromium.org
for app_browsertest.cc
BUG=585570

Review URL: https://codereview.chromium.org/1780633002

Cr-Commit-Position: refs/heads/master@{#380331}