Replace TimeTicks::UnixEpoch with Time::UnixEpoch for the net log This patch replaces the value of timeTicksOffset for the net log with a value calculated via: (Time::Now() - Time::UnixEpoch()) - (TimeTicks::Now() - TimeTicks())
In JS, TimeTicks values calculated via (t - TimeTicks()) are added to the
timeTicksOffset to calculate the wall time that they occurred.

This is equivalent to passing the TimeDelta since the epoch, along with a
reference monotonic time for subsequent TimeTicks values to subtract off.

BUG=592107

Review URL: https://codereview.chromium.org/1775663002

Cr-Commit-Position: refs/heads/master@{#380725}