Remove synthetic vsync from Android VsyncMonitor
This logic is from the ancient past and apparently is not needed anymore.

This vsync can cause BeginFrames with frame times in the past to occur.
As a part of testing https://codereview.chromium.org/1765723002, this
caused AwContentsGarbageCollectionTest#testCreateAndGcManyTimes to crash
because of this.  It would enter a situation where a synthetic vsync
would be sent that tried to predict the last vsync time, but the next
time the choreographer sent a frame time it would be *before* the
synthetic frame time.  This causes assertions later on in the pipeline
where it's assumed that all begin frames are monotonically increasing in
frame time.

These invalid begin frames used to be dropped silently by the begin
frame source multiplexer, but that logic is removed in the above patch.

R=skyostil@chromium.org,sievers@chromium.org

Review URL: https://codereview.chromium.org/1789803003

Cr-Commit-Position: refs/heads/master@{#381105}