Extensions: Extract lazy leveldb operations into LazyLevelDb.
Extracted (the bulk of) the leveldb operations from LeveldbValueStore
into a new base class: LazyLevelDb. The purpose of this change is to
support an upcoming shared database class that also needs to read/write from/to a leveldb, but partitioned for different key spaces.

This is part 2 from crrev.com/1658833003

BUG=453946

Review URL: https://codereview.chromium.org/1788893002

Cr-Commit-Position: refs/heads/master@{#381239}