[mojo-edk] Expose notification source to MojoWatch callbacks
This adds a flags argument to watch callbacks and exposes a flag
to indicate that the callback was invoked as a result of some
external process event (i.e. an incoming EDK IPC message). The
public C++ Watcher implementation is updated to take advantage
of this, effectively allowing us to dispatch to Mojo bindings
synchronously when they live on the IO thread and are servicing
messages from out-of-process.

BUG=590495

Review URL: https://codereview.chromium.org/1811433002

Cr-Commit-Position: refs/heads/master@{#381767}