Change the non-blocking event queue to the main thread event queue.
A problem with the design in that blocking touch events could get
ahead of the non-blocking events was revealed in discussions. Change
the queue so that it processes all events going to the main thread.
It may elect to put blocking events in queues that it maintains so that
events are not re-ordered when dispatched to the main thread.

Force the ack behavior in the renderer to match the disposition
of the event. The benefit of this is uncancelable touch events
now can be coalesced if the main thread is behind.

BUG=489802

Review URL: https://codereview.chromium.org/1780953003

Cr-Commit-Position: refs/heads/master@{#381821}