Remove lifecycle rewinding in FrameView::scheduleRelayout
It was added in https://codereview.chromium.org/232013002 to find extra
lifecycle violations. However, right now it is used in various cases to
schedule relayout during layout because of change of containing block chain.
This triggers the assert incorrectly (or layout should fix this, but it is
considered ok for now).

BUG=594489

Review URL: https://codereview.chromium.org/1807363002

Cr-Commit-Position: refs/heads/master@{#381997}