Dispatch SMS-schemed intent URL to default SMS handler
Previously, we were dispatching URLs that starting with sms:// to the
default SMS handler in the Android system.

In this CL, we extend it to generic intent URL that has SMS scheme such
that fallback and package name can be applied to these SMS intents as well.

BUG=584707

Review URL: https://codereview.chromium.org/1780823002

Cr-Commit-Position: refs/heads/master@{#382127}