Fix a painting bug which occurred due to the change which reduces the size of the fullscreen chrome window by 1 px on activation loss.
We reduce the size of fullscreen windows by 1px to ensure that maximized windows on the same thread don't draw over the
taskbar. This change caused a painting problem as the compositor was not aware of the changed size. This causes the compositor
to not paint correctly when the fullscreen window is activated as the window size did not change.

Setting the compositor size correctly to the window bounds size fixes this problem. The other bug I found was when the
fullscreen window is activated, we inform the delegate about the changed client size which in turn makes it across to the
webcontents. We don't want the webcontents to get notified about these size changes. Fixed by setting the flag background_fullscreen_hack_
to false after the SetBoundsInternal call.

BUG=595666

Review URL: https://codereview.chromium.org/1819633002

Cr-Commit-Position: refs/heads/master@{#382429}