Implement Chrome OS accessibility features to highlight focus, caret & cursor.
This extends the existing code that handles drawing a focus ring in an overlay
layer for ChromeVox and for the keyboard-driven OOBE, and implements support
for standalone features to highlight the focused object, the mouse cursor,
and the text caret.

The preferences to enable highlighting focus, caret, and cursor are behind
a flag, so the details of the implementation may be tweaked before this ships.
Currently each one has a distinct color, and the cursor highlight fades out
when the cursor isn't moving.

BUG=314889

Review URL: https://codereview.chromium.org/1822823002

Cr-Commit-Position: refs/heads/master@{#382693}