Make link rel=prerender always ON.
The privacy settings have the On/Off choice saying: "Use a prediction service to
load pages quickly". The Default is On.

This change allows <link rel=prerender> to perform prerendering regardless of
the setting and network type.

Reason: web developers should be able to prerender with all possible privacy
settings and with all possible network types. This would eliminate the need for
web devs to use creative ways to waste bandwidth in cases the platform is too
restrictive.

BUG=589369

Review URL: https://codereview.chromium.org/1785613009

Cr-Commit-Position: refs/heads/master@{#383064}