PlzNavigate: fix LoFi test ShouldEnableLoFiModeNavigateBackThenForward
This CL fixes
LoFiResourceDispatcherHostBrowserTest.ShouldEnableLoFiModeNavigateBackThenForward
when run with PlzNavigate enabled. The RenderFrameHost keeps track of the LoFi
state of the last successful network navigation, which allows to give the
proper value to subframes that are doing a history navigation. Subframe history
navigations do not go to the renderer when PlzNavigate is enabled, therefore
the value of LoFi state for the top level frame must be kept on the browser
side to give it to subframes.

BUG=575212
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review URL: https://codereview.chromium.org/1828663003

Cr-Commit-Position: refs/heads/master@{#383106}