Log the expected icon size when no suitable icon for a banner is found.
This CL improves the debug messaging for developers when an app banner
cannot be displayed due to an icon which is too small. The existing
messaging ("no suitable icon can be found") is split into two cases:
when all available icons are too small, and when the selected icon
cannot be fetched. In the former case, the exact required icon size
(respecting the device scale factor, which means a different sized icon
may be required on different platforms) is printed to console to better
inform developers of the icon requirements.

BUG=595517

Review URL: https://codereview.chromium.org/1826753002

Cr-Commit-Position: refs/heads/master@{#383196}