Move CompositorPendingAnimations to Oilpan's heap
CompositorPendingAnimations::m_timer can be fired in Oilpan's lazy sweeping phase.
This is problematic because at that point things accessed in the timer function
might have been already destructed.

We don't hit this problem if CompositorPendingAnimations is a GarbageCollected object
because we have logic to suppress firing timers of unreachable GarbageCollected objects.
So this CL moves CompositorPendingAnimations to the heap.

BUG=592845

Review URL: https://codereview.chromium.org/1832773002

Cr-Commit-Position: refs/heads/master@{#383248}