Fix infobar arrow tip position on MD.
On non-MD, this moves the tip down by 1 px, which basically looks better for
HTTPS and not worse for other cases.

On MD, this moves the tip down by 2 px, which makes it look perfect on non-EV
HTTPS, kinda crappy on EV, and OK everywhere else.

More importantly, it makes the tip vertical position not depend on a horizontal
layout constant, which kinda made sense in non-MD (based on how the constants
were designed there) and makes no sense at all for MD.  This is important as I'm
about to change this constant's value in a way that would otherwise look awful
on MD.

Arguably, it might make sense to position the tip differently for all the
different icons, but hardcoding a bunch of different offsets like that makes my
hackles rise.

BUG=597921
TEST=On a brand new profile (so you get the "default browser" infobar on startup), set startup page to a non-EV HTTPS site.  Start Chrome in Material Design mode.  The infobar arrow should not intersect the lock icon.

Review URL: https://codereview.chromium.org/1832023002

Cr-Commit-Position: refs/heads/master@{#383374}