Do not set m_decodeFailed in ImageFrameGenerator::decodeToYUV
This method sets m_decodeFailed bit to be true at the end on VUY
decode failure, which can result a null SkImage in

  DeferredImageDecoder::createFrameAtIndex()

when Ganesh is enabled and the image is Progressive JPEG. Ganesh
falls back to displaying an RGBA decoded image on decode failure
but the null SkImage was shown instead (see bug examples).

When decodeToYUV() fails, do not set m_decodeFailed. Instead set
m_yuvDecodingFailed to allow Ganesh fallback logic (that decodes
the image to RGBA on the CPU) to render Progressive JPEG. Update
getYUVComponentSizes() to early return if m_yuvDecodingFailed to
deny subsequent Ganesh requests to YUV decode the image.

BUG=597127

Review URL: https://codereview.chromium.org/1829923004

Cr-Commit-Position: refs/heads/master@{#383472}