Fix mini_installer in gn component builds
Due to its use of source sets the gn component mini_installer build
of setup.exe creates references to six DLLs that would not be needed if
linking with real .lib files (which avoid pulling in unneeded .obj
files). Fixing this to not pull in those DLLs would require significant
effort and would probably end up being fragile.

Therefore the pragmatic fix is to just extract the six extra DLLs. This
represents 22.4 MB of disk space (87% of that is from net.dll) which is
not enough to matter in the context of mini_installer component builds.

BUG=596885

Review URL: https://codereview.chromium.org/1835993002

Cr-Commit-Position: refs/heads/master@{#383542}