Adding test for high DPI OOPIF coordinate translation
Follow up to:
https://crrev.com/61b6c251433a39bf4e77b23df172b42df2bd1a2e

That CL applied scale factor for coordinates being received by the
browser process from an out-of-process iframe before transforming
them to the root frame's coordinate space, but did not adequately
test it. This patch adds the test.

Also, it fixes a bug that the test uncovered.

BUG=554119

Review URL: https://codereview.chromium.org/1838343002

Cr-Commit-Position: refs/heads/master@{#384321}