Fix the encoding of iso-8859-8-and-isolated*html files
These files are supposed to be in ISO-8859-8, but they're actually in
UTF-8.

./fast/text/international/iso-8859-8-and-isolate.html
./fast/text/international/iso-8859-8-and-isolate-expected.html

This CL will be landed manually because rietveld/CQ cannot handle
non-UTF-8 text files.

It's uploaded to rietveld for the record, but note that files are messed up
when viewed in rietveld.  The ground truth is what's actually in the tree
(landed manually).

Three characters in the test files are U+05D2, U+05E7, U+05DB encoded
in ISO-8859-8 (0xE2, 0xF7, 0xEB).

TBR=kojii@chromium.org
BUG=597517,559257

Review URL: https://codereview.chromium.org/1844413003 .

Cr-Commit-Position: refs/heads/master@{#384344}