Add a method to determine when to show a one-time history notice in the CBD dialog
browsing_data_ui::ShouldShowNoticeAboutOtherFormsOfBrowsingHistory
determines when we should show a permanent notice about other forms of
browsing history in the CBD dialog.

However, the one-time notice showed in a new dialog atop of CBD will
have different conditions. Add a separate method to browsing_data_ui
to determine that. Currently, the method defaults to the same value
as the aforementioned one; we are adding it to provide an interface for
the UI code. In the future, it will be updated with correct conditions.

BUG=595332

Review URL: https://codereview.chromium.org/1843003003

Cr-Commit-Position: refs/heads/master@{#384567}