Migrate chrome-layer startBrowserProcessesSync() calls to CBI.
As noted in
https://codereview.chromium.org/1809803002/#msg4,
there are several remaining locations in the chrome/
layer that still use the old startup method, which is
probably not desired.

In particular, there is at least one known crash that
this CL has been confirmed to fix (see bugs below).

BUG=560152,596871,599613

Review URL: https://codereview.chromium.org/1838953006

Cr-Commit-Position: refs/heads/master@{#384591}