Do not reset navigation state when BeforeUnload is cancelled by a commit
It is possible for a cross-site navigation to commit while a BeforeUnload
confirmation dialog is showing. This will destroy dialog and call
WebContentsImpl::OnDialogClosed. This function should not reset the navigatin
state if the RenderFrameHost that was showing the dialog is no longer the
current RenderFrameHost.

BUG=589365

Review URL: https://codereview.chromium.org/1825523002

Cr-Commit-Position: refs/heads/master@{#384674}