Add support for detecting the maximum resolution supported by the GPU for H.264 videos.
Media foundation on Windows 8.1+ supports resolutions upto 4096 x 2304. If the underlying GPU
does not support the higher resolutions, then it goes back to software decoding. While we want to
support higher H.264 resolutions for H/W decoding, we don't want to do this, if the underlying driver
does not support the same.

To detect that, we attempt to create a raw DXVA decoder device for the desired configuration and thus detect
the maximum resolution. We maintain a list of predefined resolutions, starting from 1920 x 1088 and going upto
4096 x 2304 and try and create raw devices for these resolutions.

This enables us to support resolutions of upto 4K depending on whether the GPU supports it.

BUG=573565

Review URL: https://codereview.chromium.org/1851713002

Cr-Commit-Position: refs/heads/master@{#384736}