Add keyboard accessibility to "Get help" link text on chooser UI
The "Get help" link text was a HyperlinkTextView before,
and pressing the Tab doesn't navigate to it. This patch
modified it using HyperlinkButtonCell so that pressing
the Tab can navigate to it.

This patch doesn't change the UI of the "Get help" link
text. It looks the same as before when using
HyperlinkTextView.

BUG=596671, 583479

Review URL: https://codereview.chromium.org/1835133002

Cr-Commit-Position: refs/heads/master@{#384972}