Revert ownership of the GL context objects to the ContextProvider.
Previously in https://codereview.chromium.org/1414683003 we had moved
ownership of the WebGraphicsContext3DImpl, which owns all the GL
context objects, off of ContextProvider to a skia-refcounted subclass
of GrGLInterface. This was done in order to keep them alive as long
as the GrContext was alive, which holds a ref on the GrGLInterface.

However, we can equally prevent crashes by calling abandonContext() on
the GrContext when the GL context backing it will be destroyed. So,
revert ownership back to the ContextProvider, which is simpler, and
add releaseResourcesAndAbandonContext() calls to the GrContext in
every ContextProvider's destructor.

While we're here, GrContextForWebGraphicsContext3D no longer has any
need of WebGraphicsContext3DImpl, so rename it to
GrContextForGLES2Interface and just pass it a GLES2Interface* instead.

R=piman@chromium.org
BUG=584497,551143
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1834353002

Cr-Commit-Position: refs/heads/master@{#385091}