Fix for dropping keys in Android omnibox.
On certain keyboards, you can lose a character if you
type a and [mazon.com] is added as autocomplete, and then you
quickly type something that does not match the next character
of the autocomplete text (i.e. "l").

When setting new composing text, we clear the previous autocomplete
(based on the internal span state).  The span state is invalid
after a committing the "l" above.  With this patch, we clear the
span after comitting a character that does not match the next
autocomplete character.

BUG=600875

Review URL: https://codereview.chromium.org/1865723002

Cr-Commit-Position: refs/heads/master@{#385355}