Improve <webview> accessibility on Windows.
Windows screen readers were confused by having a web document
as a child of an object with role IA2_ROLE_EMBEDDED_OBJECT.
Change it to just a GROUPING role instead.

In addition, it was possible for the child of a <webview>
to point to its child frame, but the parent of that child
woudln't point back to the same element, making it seem
orphaned and possibly breaking accessibility. Fix this by
only exposing the child of an object with a child frame when
the child frame's parent is hooked up correctly.

BUG=589325

Review URL: https://codereview.chromium.org/1862943003

Cr-Commit-Position: refs/heads/master@{#385616}