Connect LTHI::ScrollAnimated to cc::Viewport
This CL removed to hack in FindScrollLayerForDeviceViewportPoint to ensure
that we scroll the outer viewport and connects the LTHI::ScrollAnimated function
to the viewport class.

The viewport class distributes the scroll delta between the inner and outer
viewports and starts animation the layer which consumes more of the scroll delta.
Our animation system does not support running two scroll offset animations and this
is something that we may need to add in future.

BUG=575019,595421
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1866583002

Cr-Commit-Position: refs/heads/master@{#385657}