[css-grid] Removed unnecessary condition for clearing override height
In the logicalHeightForChild function we need to figure out whether the
grid item has been stretched, since in that case we had to return its
intrinsic height instead of the already computed height.

Since we had to figure that out, we did use this condition to
decide whether we should clear the flag, just in case the item had to
perform a new layout; we should avoid the override height to interfere
with the new height computation.

However, now we always return the item's logical height, so we don't
need to know whether the item has the override flag or not. Since the
cost of figuring that out is the same than clearing it in any case, it
does not make sense to have that condition.

No new tests added, because there is no change in the functionality.

Review URL: https://codereview.chromium.org/1862513008

Cr-Commit-Position: refs/heads/master@{#385707}