Sending a vote that a Password form looks like a SignUp form.
Currently for finding SignUp forms no local heuristics are used. This CL adds local heuristics for sending vote on server that a form looks similar to a SignUp form, namely if the form has more than 1 text field with the user input or it has at least 1 text field with the user input, no password and a new password field. That covers SignUp forms on many popular sites, such as gmail, amazon, wikipedia. It could give false positives on change password forms, but it's not a problem, since a decision anyway will be made server-side and there are votes for change password forms. It's considered only fields with the user input, so invisible fields are not a problem here.

BUG=582434

Review URL: https://codereview.chromium.org/1863533003

Cr-Commit-Position: refs/heads/master@{#385765}