MediaStreamRecorder: Separate video/audio origin of timestamps for WebM muxer
In certain conditions (see bug), the origin of
timestamps for audio and video is quite separated
which causes a jump on the WebM frame timestamps.
This seems to be an issue for certain players (chrome)
and not so much for others (vlc).

This CL solves the problem by keeping a different
origin-of-timestamps for audio and video separately,
and passing a TimeDelta to the AddFrame() method.

See the bug for an investigation as to why this
started happening recently.

BUG=597034
TEST=On my Win7 laptop, navigate to [1], record
stream for a few seconds, then stop recording and
try playing back. With this CL, no freeze.

[1] https://webrtc.github.io/samples/src/content/getusermedia/record/

Review URL: https://codereview.chromium.org/1863893002

Cr-Commit-Position: refs/heads/master@{#385899}