Don't force unit_tests to call TestingBrowserProcess::DeleteInstance() under obscure conditions
r327067 and r327463 forced many Chrome unit_tests to worry about browser
threads when they previously didn't have to. The errors and ways of
resolving the issue are sometimes obscure and not really anything to do
with the unit test. unit tests should be able to focus on a small area
of code and not have to add a TestBrowserThreadBundle to their test
harness if they don't need tasks invoked on background threads.

One trap is the AsyncPolicyProvider which wants to destroy parts of
itself on background threads. Provide a way for that to happen in
ChromeUnitTestSuiteInitializer::OnTestEnd() without dumping out
hard-to-diagnose stack traces onto unsuspecting test authors.

Revert the changes to first_run_bubble_unittest.cc from r327463. Another
~110 test files were also changed to ensure a BrowserThreadBundle
(rather than just a MessageLoop) is around at the appropriate time, but
first_run_bubble_unittest.cc didn't get a BrowserThreadBundle since it
would have been incompatible with the MessageLoop provided by
ViewsTestBase.

BUG=465354

Review URL: https://codereview.chromium.org/1865223002

Cr-Commit-Position: refs/heads/master@{#385928}