[Password Manager] Presave the form with generated password till the user makes successful login or removes the generated password.
To be sure we save the generated password even if we missed form submission, presave the password right after the user accepted it. Also update the password after every user input and delete it, if the user clears the password field.

The form is presaved with empty username.

BUG=582434

Review URL: https://codereview.chromium.org/1817483002

Cr-Commit-Position: refs/heads/master@{#386289}