[Media Router WebUI] Fix whitespace in filter when no devices present.
This change adds whitespace below the filter input in the filter view when there
are no devices to show for the current cast mode. In this case, the filter input
is still at the bottom of the dialog and needs to be padded as such. This change
also fixes an incorrect type annotation.

BUG=602042
R=apacible@chromium.org

Review URL: https://codereview.chromium.org/1874983005

Cr-Commit-Position: refs/heads/master@{#386627}