Looking up device id by session id for AudioRendererMixerInput
This is required for being able to correctly mix in WebRtc audio in case
of unified streams, when session id is used to specify output device
matching an open input device.

1) Modifications to AudioRendererHost and AudioOutputDevice to retrieve
the matched device id in OnRequestDeviceAuthorization()/OnDeviceAuthorized()

-- This is only manually tested so far. I tried to come up with a neat unit
test but AudioRendererHost is lacking mocks (uses "for test" settings of
its dependencies instead) so I got lost there. As much as I want to have the
unit test in place, I would prefer to not spend any more time on this now.
The change is pretty straight forward. Please let me know if you think
the unit test is a must-have.

2) AudioRendereMixer key (in AudioRendererMixerManager) is modified
to take into account default device specifics.

BUG=587457

Review URL: https://codereview.chromium.org/1769933002

Cr-Commit-Position: refs/heads/master@{#386652}