[WebView] Disallow geolocation on insecure origins for apps targeting N and higher.
In Chrome geolocation has been disabled on insecure origins
previously in crrev.com/1485973002.

For WebView we are enabling same behavior for apps that are
targeting N and higher versions of Android. This is to allow
for a smooth transition and remain reasonably backwards
compatible with older apps.

BUG=520765

Review URL: https://codereview.chromium.org/1882783002

Cr-Commit-Position: refs/heads/master@{#387299}