Fire accessibility events to connect a child frame to its parent.
Previous changes correctly hooked up accessibility trees for
out-of-process iframes and guest views (like <webview>) to their
parent accessibility tree. However, on Windows we need to do one
additional step and fire events on the parent node (both
a REORDER event and IA2_EVENT_TEXT_INSERTED) so that AT like
NVDA can pick up the change to the virtual buffer.

To make this robust, this change adds a new boolean that keeps
track of whether a BrowserAccessibilityManager is attached to
its parent or not. It gets checked after every event and reset
if the tree ID or the root of the tree changes.

BUG=589325

Review URL: https://codereview.chromium.org/1880303002

Cr-Commit-Position: refs/heads/master@{#387712}