[OnionSoup] Moving WakeLock Service to Blink (2/2): removing WebWakeLockClient and WakeLockDispatcher
This CL is part 2/2 of moving the WakeLockService renderer parts
from content/ to WebKit/. The complete transfer can be found in
https://crrev.com/1794553002, of which this is the part removing
the WebWakeLockClient and its renderer implementation. See bug
for more info.

This CL removes the files
 - content/renderer/wake_lock/wake_lock_dispatcher.*
 - WebKit/Source/platform/wake_lock/WebWakeLockClient.h

since that functionality is folded into
  WebKit/public/platform/modules/wake_lock/ScreenWakeLock.*
of which some methods are made private.

modified to talk directly to the mojo service impl in
browser's render frame host.

The WebKit unittests ScreenWakeLockTest.cpp:
- is refactored to use a MockServiceRegistry and a
mojo MockWakeLockService.
- is NOT moved out of web/ because it needs a couple
of Test infra located there. This unit test should be
made into a (series of) LayoutTests, but that's beyond this CL.

BUG=600895

TESTS= in particular:
./out/Debug/webkit_unit_tests  --gtest_filter=*ScreenWakeLockTest.*
./out/Debug/content_browsertests --gtest_filter=WakeLockTest.*

Review URL: https://codereview.chromium.org/1883493003

Cr-Commit-Position: refs/heads/master@{#387786}