Fix size_t truncation warnings in gfx
The plane parameter to some gfx functions was logically being treated
as a size_t in many places and was declared as a size_t parameter in
one function, but int elsewhere. This makes it consistently size_t in
order to avoid C4267 (size_t to smaller truncation) warnings without
adding any casts.

Note that in 64-bit processes size_t may be more efficient than int
and should be no less.

The change of RowSizeForBufferFormat's plane argument to size_t was done
for consistency, not to suppress a particular warning.

BUG=566113

Review URL: https://codereview.chromium.org/1862793004

Cr-Commit-Position: refs/heads/master@{#387817}