Introduce SpdyFramer::SPDY_INVALID_STREAM_ID.
A stream ID of 0 is required/forbidden by the spec depending on the frame type.
Catch these failures in the framer as a SPDY_INVALID_STREAM_ID error.  Also
attempt to centralize some of the validation logic in the framer.

This CL lands server changes 117147997 and 117463896 by dahollings.

BUG=488484

Review URL: https://codereview.chromium.org/1891743002

Cr-Commit-Position: refs/heads/master@{#388220}