Provide web contents to Chrome's mojo services
For the mojo service implementations in Chrome that fulfill tasks for
render frames, the ChromeServiceRegistrar should provide the WebContents
object that corresponds to the render frame that requested the service.
This allows the service implementation to access WebContents-specific
data and show UI on top of the WebContents.

This patch also renames PaymentRequestDialog.java into
PaymentRequestImpl.java, because it implements the mojo PaymentRequest
service, while the UI functionality is implemented in a separate class.

BUG=587995

Review URL: https://codereview.chromium.org/1902913002

Cr-Commit-Position: refs/heads/master@{#388310}