Set site engagement timestamps to privacy-respectful values when history is cleared.
The site engagement service stores last updated and last shortcut launch
timestamps per origin. When history is cleared, the former is set to the
time of history clearing and the latter is not changed.

This CL sets both timestamps to the time of the most recent remaining
visit in history for that origin once history clearing is completed. The
last shortcut launch time is only set if it was non-null prior to the
history being cleared.

A slight change to the history service query (adding the last visited date
to the data returned) is necessary to enable this change. More
comprehensive testing of the history expiry mechanism are added to
ensure correctness.

BUG=604284

Review URL: https://codereview.chromium.org/1901563002

Cr-Commit-Position: refs/heads/master@{#388321}