Replicate across OOPIFs flags for dumping of various WebFrameClient callbacks.
The replicated flags mean that in some cases test output includes
"duplicated" calls to WebFrameClient::willSendRequest and/or
WebFrameClient::didFinishResourceLoad methods.  This is expected [1]
without PlzNavigate.  Replication of the test flags is still desirable,
so to make the tests behave the same with and without --site-per-process
we suppress dumping of these kinds of callbacks - we can do this,
because verification of the presence of this callbacks was not essential
to the core verification done by the tests (that cross-site access is
blocked).

[1] https://groups.google.com/a/chromium.org/d/topic/site-isolation-dev/XRYuFNi4Px0/discussion

BUG=587175

Review URL: https://codereview.chromium.org/1895543002

Cr-Commit-Position: refs/heads/master@{#388512}