Remove "null" InputTypeView.
The "null" InputTypeView was used in the case of an INPUT element with author
ShadowRoot to represent a view without any form control behavior. We can't add
author ShadowRoot to INPUT elements any longer, and the "null" InputTypeView is
unnecessary now.

Since this CL, InputTypeView class switches from the "null" InputTypeView to the
most common InputTypeView implementation for all of InputTypes.  So we move
override functions in InputType to InputTypeView.

This CL has no behavior changes.

BUG=243714

Review URL: https://codereview.chromium.org/1907883002

Cr-Commit-Position: refs/heads/master@{#388715}