Initialize Material Design in views_examples_with_content_exe
This is required since r386582 requires MaterialDesignController
initialization to be explicit.

It was fixed for views_examples in r388338

This applies a similar fix for views_content_client which is used by
views_examples_with_content_exe

Also, add an OWNERS file to ui/views_content_client with tapted@.
Probably should have done this when I made it in r272902 but I was too
timid.

BUG=605126

Review URL: https://codereview.chromium.org/1903393003

Cr-Commit-Position: refs/heads/master@{#389037}