Ignore LNK4197 warning for osmesa.dll
This warning was diagnosed as benign in http://crbug.com/512956 and it
thus shouldn't spam every dev's build output.
If we otherwise think it should be addressed, the bug should be kept
opened and assigned but having every dev see this warning is not useful.

BUG=512956

Review URL: https://codereview.chromium.org/1910823002

Cr-Commit-Position: refs/heads/master@{#389185}