SelectionOwner: add support for TIMESTAMP target
Even though XSetSelectionOwner() has a Time argument, there's no way to
retrieve the timestamp associated with the selection from the X server
itself. Instead, the X11 selection protocol requires selections to
provide a TIMESTAMP target.

Unfortunately, there's also no good way to get the current timestamp
from the X server: one way is to make a no-op property change and
observe the resulting PropertyNotify event. This is the what GTK does,
but it both blocks and performs out-of-order event dispatching.

In Chromium, we control our event source, so we implement an alternate
strategy instead: we record the server time seen in X events with a
time field and use the last seen value when taking ownership of the X
selection. Of course, since this is X, the server time observer has to
be careful to ignore instances where time == CurrentTime (explicitly
set by some programs to indicate an X event is synthesized) and where
time goes backwards (since events can be reordered).

Note that despite all this logic to sanitize X server time, actually
using the last seen time when calling XSetSelectionOwner() doesn't
actually seem to be good enough. Random omnibox tests still fail in
that case, so we still pass CurrentTime to XSetSelectionOwner() to
increase the chances of success.

¯\_(ツ)_/¯

BUG=395077

Review URL: https://codereview.chromium.org/1892483005

Cr-Commit-Position: refs/heads/master@{#389208}