[css-grid] Implement auto-repeat computation
We added support for parsing and style for the new auto-repeat syntax. This
CL actually implements the feature by computing the number of auto-repeat
tracks that will be part of the explicit grid depending on the available
size on the corresponding axis.

Note that this CL does not drop the empty tracks, i.e., auto-fit will work
exactly as auto-fill until the empty track removal is implemented.

Test results are not totally correct yet because we need to add the line
names to the computed style. As that requires a rather large piece of code
it will be done in a follow up patch. Track sizes are correct though.

BUG=589460

Review URL: https://codereview.chromium.org/1895443002

Cr-Commit-Position: refs/heads/master@{#389456}